requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
marshmallow
python-dotenv
Flask-Admin 